Change implementation of tensor.can_cast, tensor.result_type
Implementation uses device-specific type promotion rules if input
are arrays.

For result_type, all arrays must have the same dtype support, or
an exception is raused.
